UPSTREAM: KVM: arm64: Use less bits for hyp_page order
The hyp_page order is currently encoded on 4 bytes even though it is
guaranteed to be smaller than this. Make it 2 bytes to reduce the hyp
vmemmap overhead.

Link: https://lore.kernel.org/r/20210608114518.748712-7-qperret@google.com
(cherry picked from commit 87ec0606733e1aa9568f54ddb41f03aa6b5687f2)
Bug: 192636784

+++ arch/arm64/kvm/hyp/include/nvhe/memory.h
@@ -9,7 +9,7 @@
 
 struct hyp_page {
 	unsigned int refcount;
-	unsigned int order;
+	unsigned short order;
 };
 
 extern s64 hyp_physvirt_offset;
